+++ src/fill_lmdb_plugin.cpp
@@ -16,6 +16,7 @@ using namespace std::literals;
 
 namespace asio      = boost::asio;
 namespace bpo       = boost::program_options;
+namespace kv        = state_history::kv;
 namespace lmdb      = state_history::lmdb;
 namespace websocket = boost::beast::websocket;
 
@@ -30,7 +31,7 @@ struct lmdb_table;
 struct lmdb_field {
     std::string                 name        = {};
     const abieos::abi_field*    abi_field   = {};
-    const lmdb::type*           type        = {};
+    const kv::type*             type        = {};
     std::unique_ptr<lmdb_table> array_of    = {};
     std::unique_ptr<lmdb_table> optional_of = {};
     abieos::input_buffer        pos         = {}; // temporary filled by fill()
@@ -111,17 +112,17 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
         ilog("verifying expected records are present");
         uint32_t expected = first;
-        for_each_subkey(t, lmdb_inst->db, lmdb::make_block_key(0), lmdb::make_block_key(0xffff'ffff), [&](auto&, auto k, auto v) {
+        for_each_subkey(t, lmdb_inst->db, kv::make_block_key(0), kv::make_block_key(0xffff'ffff), [&](auto&, auto k, auto v) {
             auto orig_k = k;
-            if (lmdb::bin_to_key_tag(k) != lmdb::key_tag::block)
+            if (kv::bin_to_key_tag(k) != kv::key_tag::block)
                 throw std::runtime_error("This shouldn't happen (1)");
-            auto block_num = lmdb::bin_to_native_key<uint32_t>(k);
-            auto tag       = lmdb::bin_to_key_tag(k);
-            if (tag == lmdb::key_tag::table_row && (block_num < first || block_num > head))
+            auto block_num = kv::bin_to_native_key<uint32_t>(k);
+            auto tag       = kv::bin_to_key_tag(k);
+            if (tag == kv::key_tag::table_row && (block_num < first || block_num > head))
                 throw std::runtime_error(
                     "Saw row for block_num " + std::to_string(block_num) +
-                    ", which is out of range [first, head]. key: " + lmdb::key_to_string(orig_k));
-            if (tag != lmdb::key_tag::received_block)
+                    ", which is out of range [first, head]. key: " + kv::key_to_string(orig_k));
+            if (tag != kv::key_tag::received_block)
                 return true;
             if (block_num == first || block_num == head || !(block_num % 10'000))
                 ilog("Found records for block ${b}", ("b", block_num));
@@ -137,7 +138,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
         ilog("verifying table_index keys reference existing records");
         uint32_t num_ti_keys = 0;
-        for_each(t, lmdb_inst->db, lmdb::make_table_index_key(), lmdb::make_table_index_key(), [&](auto k, auto v) {
+        for_each(t, lmdb_inst->db, kv::make_table_index_key(), kv::make_table_index_key(), [&](auto k, auto v) {
             if (!((++num_ti_keys) % 1'000'000))
                 ilog("Checked ${n} table_index keys", ("n", num_ti_keys));
             if (!lmdb::exists(t, lmdb_inst->db, lmdb::to_const_val(v)))
@@ -148,7 +149,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
         ilog("verifying table_index_ref keys reference existing table_index records");
         uint32_t num_ti_ref_keys = 0;
-        for_each(t, lmdb_inst->db, lmdb::make_table_index_ref_key(), lmdb::make_table_index_ref_key(), [&](auto k, auto v) {
+        for_each(t, lmdb_inst->db, kv::make_table_index_ref_key(), kv::make_table_index_ref_key(), [&](auto k, auto v) {
             if (!((++num_ti_ref_keys) % 1'000'000))
                 ilog("Checked ${n} table_index_ref keys", ("n", num_ti_ref_keys));
             if (!lmdb::exists(t, lmdb_inst->db, lmdb::to_const_val(v)))
@@ -215,8 +216,8 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
                 raw_type = raw_type->optional_of;
             if (raw_type->array_of)
                 raw_type = raw_type->array_of;
-            auto type_it = lmdb::abi_type_to_lmdb_type.find(raw_type->name);
-            if (type_it == lmdb::abi_type_to_lmdb_type.end() && !array_of_struct && !array_of_variant && !optional_of_struct)
+            auto type_it = kv::abi_type_to_kv_type.find(raw_type->name);
+            if (type_it == kv::abi_type_to_kv_type.end() && !array_of_struct && !array_of_variant && !optional_of_struct)
                 throw std::runtime_error("don't know lmdb type for abi type: " + raw_type->name);
 
             table.fields.push_back(std::make_unique<lmdb_field>());
@@ -244,8 +245,8 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     }
 
     void add_table(const std::string& table_name, const std::string& table_type, const jarray& key_names) {
-        auto table_name_it = lmdb::table_names.find(table_name);
-        if (table_name_it == lmdb::table_names.end())
+        auto table_name_it = kv::table_names.find(table_name);
+        if (table_name_it == kv::table_names.end())
             throw std::runtime_error("unknown table \"" + table_name + "\"");
         if (tables.find(table_name) != tables.end())
             throw std::runtime_error("duplicate table \"" + table_name + "\"");
@@ -368,7 +369,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     }
 
     void load_fill_status(lmdb::transaction& t) {
-        current_db_status = lmdb::get<state_history::fill_status>(t, lmdb_inst->db, lmdb::make_fill_status_key(), false);
+        current_db_status = lmdb::get<state_history::fill_status>(t, lmdb_inst->db, kv::make_fill_status_key(), false);
         if (!current_db_status)
             return;
         head            = current_db_status->head;
@@ -379,7 +380,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     }
 
     void check_conflicts(lmdb::transaction& t) {
-        auto r = lmdb::get<state_history::fill_status>(t, lmdb_inst->db, lmdb::make_fill_status_key(), false);
+        auto r = lmdb::get<state_history::fill_status>(t, lmdb_inst->db, kv::make_fill_status_key(), false);
         if ((bool)r != (bool)current_db_status || (r && *r != *current_db_status))
             throw std::runtime_error("Another process is filling this database");
     }
@@ -388,7 +389,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
         jarray result;
         if (head) {
             for (uint32_t i = irreversible; i <= head; ++i) {
-                auto rb = lmdb::get<lmdb::received_block>(t, lmdb_inst->db, lmdb::make_received_block_key(i));
+                auto rb = lmdb::get<kv::received_block>(t, lmdb_inst->db, kv::make_received_block_key(i));
                 result.push_back(jvalue{jobject{
                     {{"block_num"s}, jvalue{std::to_string(rb->block_num)}},
                     {{"block_id"s}, jvalue{(std::string)rb->block_id}},
@@ -405,22 +406,22 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
         else
             current_db_status = state_history::fill_status{
                 .head = head, .head_id = head_id, .irreversible = head, .irreversible_id = head_id, .first = first};
-        put(t, lmdb_inst->db, lmdb::make_fill_status_key(), *current_db_status, true);
+        put(t, lmdb_inst->db, kv::make_fill_status_key(), *current_db_status, true);
     }
 
     void truncate(lmdb::transaction& t, uint32_t block) {
-        for_each(t, lmdb_inst->db, lmdb::make_block_key(block), lmdb::make_block_key(), [&](auto k, auto v) {
+        for_each(t, lmdb_inst->db, kv::make_block_key(block), kv::make_block_key(), [&](auto k, auto v) {
             lmdb::check(mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(k)), nullptr), "truncate (1): ");
             return true;
         });
-        for_each(t, lmdb_inst->db, lmdb::make_table_index_ref_key(block), lmdb::make_table_index_ref_key(), [&](auto k, auto v) {
+        for_each(t, lmdb_inst->db, kv::make_table_index_ref_key(block), kv::make_table_index_ref_key(), [&](auto k, auto v) {
             std::vector<char> k2{k.pos, k.end};
             lmdb::check(mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(v)), nullptr), "truncate (2): ");
             lmdb::check(mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(k2)), nullptr), "truncate (3): ");
             return true;
         });
 
-        auto rb = lmdb::get<lmdb::received_block>(t, lmdb_inst->db, lmdb::make_received_block_key(block - 1), false);
+        auto rb = lmdb::get<kv::received_block>(t, lmdb_inst->db, kv::make_received_block_key(block - 1), false);
         if (!rb) {
             head    = 0;
             head_id = {};
@@ -480,8 +481,8 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
                 first = head;
             write_fill_status(*active_tx);
 
-            put(*active_tx, lmdb_inst->db, lmdb::make_received_block_key(result.this_block->block_num),
-                lmdb::received_block{result.this_block->block_num, result.this_block->block_id});
+            put(*active_tx, lmdb_inst->db, kv::make_received_block_key(result.this_block->block_num),
+                kv::received_block{result.this_block->block_num, result.this_block->block_id});
 
             if (commit_now) {
                 if (config->enable_trim) {
@@ -544,7 +545,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     void receive_block(uint32_t block_num, const checksum256& block_id, input_buffer bin, lmdb::transaction& t) {
         state_history::signed_block block;
         bin_to_native(block, bin);
-        auto              key = lmdb::make_block_info_key(block_num);
+        auto              key = kv::make_block_info_key(block_num);
         std::vector<char> value;
 
         std::vector<uint32_t> positions;
@@ -571,10 +572,10 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
         std::vector<char> index_key;
         for (auto& [_, index] : block_info_table->indexes) {
             index_key.clear();
-            lmdb::append_table_index_key(index_key, block_info_table->short_name, index.name);
+            kv::append_table_index_key(index_key, block_info_table->short_name, index.name);
             fill_key(index_key, index);
             lmdb::put(t, lmdb_inst->db, index_key, key);
-            lmdb::put(t, lmdb_inst->db, lmdb::make_table_index_ref_key(block_num, key, index_key), index_key);
+            lmdb::put(t, lmdb_inst->db, kv::make_table_index_ref_key(block_num, key, index_key), index_key);
         }
     } // receive_block
 
@@ -603,7 +604,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
                         ("b", block_num)("t", table_delta.name)("n", num_processed)("r", table_delta.rows.size()));
                 check_variant(row.data, *table.abi_type, 0u);
                 delta_key.clear();
-                lmdb::append_delta_key(delta_key, block_num, row.present, table.short_name);
+                kv::append_delta_key(delta_key, block_num, row.present, table.short_name);
                 value.clear();
                 abieos::native_to_bin(value, block_num);
                 abieos::native_to_bin(value, row.present);
@@ -614,11 +615,11 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
                 for (auto& [_, index] : table.indexes) {
                     index_key.clear();
-                    lmdb::append_table_index_key(index_key, table.short_name, index.name);
+                    kv::append_table_index_key(index_key, table.short_name, index.name);
                     fill_key(index_key, index);
-                    lmdb::append_table_index_state_suffix(index_key, block_num, row.present);
+                    kv::append_table_index_state_suffix(index_key, block_num, row.present);
                     lmdb::put(t, lmdb_inst->db, index_key, delta_key);
-                    lmdb::put(t, lmdb_inst->db, lmdb::make_table_index_ref_key(block_num, delta_key, index_key), index_key);
+                    lmdb::put(t, lmdb_inst->db, kv::make_table_index_ref_key(block_num, delta_key, index_key), index_key);
                 }
                 ++num_processed;
             }
@@ -645,7 +646,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
         uint32_t transaction_ordinal = ++num_ordinals;
 
         std::vector<char> key;
-        lmdb::append_transaction_trace_key(key, block_num, ttrace.id);
+        kv::append_transaction_trace_key(key, block_num, ttrace.id);
 
         std::vector<char> value;
         abieos::native_to_bin(value, block_num);
@@ -678,7 +679,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
         const state_history::action_trace_v0& atrace, std::vector<char>& key, std::vector<char>& value, std::vector<char>& index_key) {
 
         key.clear();
-        lmdb::append_action_trace_key(key, block_num, ttrace.id, atrace.action_ordinal.value);
+        kv::append_action_trace_key(key, block_num, ttrace.id, atrace.action_ordinal.value);
         value.clear();
 
         std::vector<uint32_t> positions;
@@ -721,10 +722,10 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
         for (auto& [_, index] : action_trace_table->indexes) {
             index_key.clear();
-            lmdb::append_table_index_key(index_key, action_trace_table->short_name, index.name);
+            kv::append_table_index_key(index_key, action_trace_table->short_name, index.name);
             fill_key(index_key, index);
             lmdb::put(t, lmdb_inst->db, index_key, key);
-            lmdb::put(t, lmdb_inst->db, lmdb::make_table_index_ref_key(block_num, key, index_key), index_key);
+            lmdb::put(t, lmdb_inst->db, kv::make_table_index_ref_key(block_num, key, index_key), index_key);
         }
 
         // todo: receipt_auth_sequence
@@ -734,10 +735,10 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
     template <typename F>
     void for_each_row_in_block(lmdb::transaction& t, uint32_t block_num, F f) {
-        auto row_bound = lmdb::make_table_row_key(block_num);
+        auto row_bound = kv::make_table_row_key(block_num);
         lmdb::for_each(t, lmdb_inst->db, row_bound, row_bound, [&](auto k, auto row_content) {
             k.pos += row_bound.size();
-            auto table_name = lmdb::bin_to_native_key<abieos::name>(k);
+            auto table_name = kv::bin_to_native_key<abieos::name>(k);
             auto pk         = k;
             return f(table_name, pk);
         });
@@ -746,11 +747,11 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     template <typename F>
     void for_each_delta_in_block(lmdb::transaction& t, uint32_t block_num, F f) {
         std::vector<char> delta_bound;
-        lmdb::append_delta_key(delta_bound, block_num);
+        kv::append_delta_key(delta_bound, block_num);
         lmdb::for_each(t, lmdb_inst->db, delta_bound, delta_bound, [&](auto k, auto row_content) {
             k.pos += delta_bound.size();
-            auto table_name = lmdb::bin_to_native_key<abieos::name>(k);
-            auto present    = lmdb::bin_to_native_key<bool>(k);
+            auto table_name = kv::bin_to_native_key<abieos::name>(k);
+            auto present    = kv::bin_to_native_key<bool>(k);
             auto pk         = k;
             return f(table_name, present, pk);
         });
@@ -759,29 +760,29 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
     template <typename F>
     void for_each_row_trim(lmdb::transaction& t, abieos::name table_name, abieos::input_buffer pk, uint32_t block_num, F f) {
         std::vector<char> trim_bound;
-        lmdb::append_table_index_key(trim_bound, table_name, "trim"_n);
-        lmdb::native_to_bin_key<uint32_t>(trim_bound, block_num);
+        kv::append_table_index_key(trim_bound, table_name, "trim"_n);
+        kv::native_to_bin_key<uint32_t>(trim_bound, block_num);
         lmdb::for_each(t, lmdb_inst->db, trim_bound, trim_bound, [&](auto k, auto v) { return f(v); });
     }
 
     template <typename F>
     void for_each_delta_trim(lmdb::transaction& t, abieos::name table_name, abieos::input_buffer pk, uint32_t max_block_num, F f) {
         std::vector<char> trim_bound;
-        lmdb::append_table_index_key(trim_bound, table_name, "trim"_n);
+        kv::append_table_index_key(trim_bound, table_name, "trim"_n);
         trim_bound.insert(trim_bound.end(), pk.pos, pk.end);
         auto trim_lower_bound = trim_bound;
-        lmdb::native_to_bin_key<uint32_t>(trim_lower_bound, ~max_block_num);
+        kv::native_to_bin_key<uint32_t>(trim_lower_bound, ~max_block_num);
 
         lmdb::for_each(t, lmdb_inst->db, trim_lower_bound, trim_bound, [&](auto k, auto v) {
             k.pos += trim_bound.size();
-            auto block_num = ~lmdb::bin_to_native_key<uint32_t>(k);
-            auto present   = !lmdb::bin_to_native_key<bool>(k);
+            auto block_num = ~kv::bin_to_native_key<uint32_t>(k);
+            auto present   = !kv::bin_to_native_key<bool>(k);
             return f(block_num, present);
         });
     }
 
     void remove_row(lmdb::transaction& t, abieos::name table_name, uint32_t block_num, std::vector<char> key) {
-        auto index_ref_bounds = lmdb::make_table_index_ref_key(block_num, key);
+        auto index_ref_bounds = kv::make_table_index_ref_key(block_num, key);
         lmdb::for_each(t, lmdb_inst->db, index_ref_bounds, index_ref_bounds, [&](auto k, auto v) {
             lmdb::check(mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(v)), nullptr), "remove_row (1): ");
             lmdb::check(mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(k)), nullptr), "remove_row (2): ");
@@ -792,7 +793,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
 
     void remove_delta(lmdb::transaction& t, abieos::name table_name, uint32_t block_num, bool present, abieos::input_buffer pk) {
         std::vector<char> delta_key;
-        lmdb::append_delta_key(delta_key, block_num, present, table_name);
+        kv::append_delta_key(delta_key, block_num, present, table_name);
         delta_key.insert(delta_key.end(), pk.pos, pk.end);
         remove_row(t, table_name, block_num, delta_key);
     }
@@ -833,8 +834,7 @@ struct flm_session : std::enable_shared_from_this<flm_session> {
                 return true;
             });
             lmdb::check(
-                mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(lmdb::make_received_block_key(block_num))), nullptr),
-                "trim: ");
+                mdb_del(t.tx, lmdb_inst->db.db, lmdb::addr(lmdb::to_const_val(kv::make_received_block_key(block_num))), nullptr), "trim: ");
         }
 
         ilog("trim: removed ${r} rows and ${d} deltas", ("r", num_rows_removed)("d", num_deltas_removed));
